Move new API classes & RESTRequest into own files
